+++ src/argtable/arg_dbl.c
@@ -133,8 +133,8 @@ struct arg_dbl* arg_dbl1(const char* shortopts,
 struct arg_dbl* arg_dbln(const char* shortopts,
                                const char* longopts,
                                const char *datatype,
-                               int mincount,
-                               int maxcount,
+                               size_t mincount,
+                               size_t maxcount,
                                const char *glossary)
     {
     size_t nbytes;
